requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
accelerate>=0.20.0
bitsandbytes>=0.39.0

# LLaVA-Med Dependencies
llava-med>=1.5.0
Pillow>=10.0.0
requests>=2.31.0

# Quantum Computing
qiskit>=0.45.0
qiskit-aer>=0.12.0
qiskit-machine-learning>=0.7.0
pennylane>=0.32.0
pennylane-qiskit>=0.32.0

# Web Framework
Flask>=2.3.0
Flask-CORS>=4.0.0
Flask-JWT-Extended>=4.5.0
gunicorn>=21.0.0

# AI/ML Libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Explainable AI
lime>=0.2.0.1
shap>=0.42.0
captum>=0.6.0
grad-cam>=1.4.8

# API Integration
openai>=1.0.0
python-dotenv>=1.0.0

# Report Generation
reportlab>=4.0.0
jinja2>=3.1.0
weasyprint>=59.0

# Data Processing
nibabel>=5.1.0  # For medical imaging formats
pydicom>=2.4.0  # For DICOM files
SimpleITK>=2.2.0  # For medical image processing

# Security & Compliance
cryptography>=41.0.0
pycryptodome>=3.18.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0